separate out where, mask tets
